ex02

+++ module_09/ex02/PmergeMe.hpp
@@ -15,7 +15,7 @@ private:
     std::vector<unsigned int> vec;
     std::deque<unsigned int> deq;
     std::vector<unsigned int> _j_seq; // for index
-    std::vector<unsigned int> jacobstal_seq(size_t size);
+    std::vector<unsigned int> jacobstal_seq(unsigned int size);
     unsigned int max; // for insertion, left is max, second in pair
 
     unsigned int left;  // jac sequence _j_seq[i]
@@ -36,7 +36,7 @@ public:
     std::vector<unsigned int> getVec() const;
     void printVec(std::string str) const;
     void printDeq(std::string str) const;
-    // void sort();
+    //  void sort();
 
     void PmergeMeVec();
     void pmergeDeq();
